[FIX] Delete two useless parameters
The `gradient_approximation` and `testSampleSkip` parameters are used to
set whether to adopt the relevant ML speedup technology, which should be
implemented by the user, so delete the parameter.

In addition, in the MC calculation function, the judgment of whether to
perform exact calculation or MC sampling calculation has been modified.